requirements.txt
Django>=4.2,<5.0
Pillow>=10.0
python-decouple>=3.8
boto3>=1.28          # Pour AWS S3 (stockage cloud des photos)
django-storages>=1.14 # Backend S3/local configurable
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
whitenoise>=6.6      # Servir les fichiers statiques en production
psycopg2-binary>=2.9 # PostgreSQL (optionnel, sqlite par défaut)
